remove play funcs
