+++ game/scripts/vscripts/events.lua
@@ -41,6 +41,7 @@ function AIGameMode:OnGameStateChanged(keys)
             self:PreGameOptions()
         end
     elseif state == DOTA_GAMERULES_STATE_PRE_GAME then
+        -- TODO 需要初始化，移植到TS中
         local gameDifficulty = CustomNetTables:GetTableValue('game_difficulty', 'all').difficulty
         -- modifier towers
         local tTowers = Entities:FindAllByClassname("npc_dota_tower")
@@ -213,23 +214,6 @@ function AIGameMode:RefreshGameStatus()
     AIGameMode.creepBuffLevelMegaBad = buffLevelMegaBad
 end
 
--- function AIGameMode:OnLastHit(keys)
---     if keys.FirstBlood == 1 then
---         local hero = PlayerResource:GetSelectedHeroEntity(keys.PlayerID)
---         if hero and hero:HasAbility("Hero_vo_player") then
---             hero:PlayVoiceAllPlayerIgnoreCooldown(hero:GetName() .. ".vo.FirstBlood")
---         end
---     end
--- end
-
--- function AIGameMode:OnPickHeroSpawn(keys)
---     local heroname = keys.hero
---     local hero = EntIndexToHScript(keys.heroindex)
---     if hero:HasAbility("Hero_vo_player") then
---         hero:PlayVoiceIgnoreCooldown(heroname .. ".vo.Spawn")
---     end
--- end
-
 function AIGameMode:OnNPCSpawned(keys)
     if GameRules:State_Get() < DOTA_GAMERULES_STATE_PRE_GAME then
         Timers:CreateTimer(1, function()
@@ -240,19 +224,6 @@ function AIGameMode:OnNPCSpawned(keys)
     local hEntity = EntIndexToHScript(keys.entindex)
     if not hEntity or hEntity:IsNull() then
         return
-    end
-
-    if hEntity:IsBaseNPC() then
-        local playerid = hEntity:GetPlayerOwnerID()
-        if playerid then
-            local hero = PlayerResource:GetSelectedHeroEntity(playerid)
-            if hero and hero == hEntity and hero:HasAbility("Hero_vo_player") then
-                if not hero.isBuyBack then
-                    hero:PlayVoiceIgnoreCooldown(hero:GetName() .. ".vo.Respawn")
-                end
-                hero.isBuyBack = false
-            end
-        end
     end
 
     local sName = hEntity:GetName()
